Add remove job method to QueueManager (#2205)
https://eaflood.atlassian.net/browse/WATER-4055

The need for this was discovered when an Annual bill run became stuck in 'SENDING'. The service failed with a timeout whilst firing the `/approve` and `/send` requests to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). The fact the CHA responded with `204` to both requests immediately is a thing to look into another day!

But WRLS thought it got a timeout and then blew up. When we tested we found the first time this happens the bill run is sent back to the `READY` status. But when you try again the bill run becomes stuck in `SENDING`.

Our observation was that the controller updates the billing batch's status to `SENDING` after adding the 'approve-batch' job to the queue. Only nothing then happened. The reason is that the previous 'job' still appears to be in the queue, though in a failed state. Because `src/modules/billing/jobs/approve-batch.js` uses the billing batch ID for the job ID, BullMQ is seeing the second entry as a duplicate and ignores it.

Hence batch gets set to `SENDING`, no attempt to approve the bill run actually takes place, and the UI spins forever more.

We're actively trying to move away from message queues like BullMQ and PGBoss because of this kind of added complexity. If you don't properly understand the frameworks you get yourself into this kind of pickle, which this service is riddled with!

So, this change adds support for removing a specific job to `QueueManager`. We're just calling [BullMQ's remove()](https://api.docs.bullmq.io/classes/v4.Queue.html#remove) behind the scenes. But it solves this specific scenario and might prove useful should we find any others.