Fix stuck Midlands sending bill run
https://eaflood.atlassian.net/browse/WATER-4055

On July 7 2023 the B&D team 'sent' the Midlands annual bill run. It then appeared to get stuck in that state.

A check of the logs confirmed the Charging Module received the 'approve' and 'send' requests successfully and that it
completed the bill (marking it as BILLED and therefore no longer possible to make changes).

This means the bill run data in WRLS is out of sync and needs to be updated.

We restored a backup into pre-prod and exported the water.billing_batches, billing_invoices, billing_invoice_licences
and billing_transactions records for the bill run. We then ran an edited version of the water-abstraction-service with
the approve and send requests commented out. That meant the WRLS service could then carry on to the remaining steps
it failed to complete previously.

With that done we then exported the same records again and compared the results. This fix migration is the result. It
updates those fields which were different in the DB after the bill run was sent.

Due to the size we would have preferred that this was NOT a migration. But our outdated RfC process meant we could
get this fix applied far quicker if it was.