Fix `update-invoices-worker` for sroc reissuing (#2213)
https://eaflood.atlassian.net/browse/WATER-4036

When `update-invoices-worker` handles sroc transactions, it maps the CM transaction to a billing transaction then updates some sroc-specific fields. This fails when handlng transactions created during reissuing.

The problem lies in the way that it determines the transaction's `billingTransactionId`: it takes the field `clientId` on the CM transaction (ie. our own billing transaction id), which is fine when we have sent a "create transaction" request across, as we include this in the request. However if the transaction was created by the CM itself during reissuing, it won't have a client id and therefore we would be trying to work with a transaction that has no transaction id!

The fix for this is straightforward. Our `mapTransaction` function already receives a map of all our transactions on the invoice, indexed by CM transaction id. We can therefore take the CM transaction we're mapping and use its id to pull our transaction from the map of transactions, which then lets us get our transaction's id.

In theory we could always do this; however for the sake of performance we only look this up if the client id isn't specific in the CM transaction.