Fix stuck annual Anglian bill run
https://eaflood.atlassian.net/browse/WATER-4074

An annual bill run has become 'stuck' during the send process. But the step in the process that let the Charging Module know succeeded. So, now the bill run is complete in the CHA and cannot be changed but reset to **READY** in WRLS.

This happened before in WATER-4055. Until we get to the bottom of the issue the way we rectify it is to deploy a version of the service that skips dealing with the CHA during the send process to PRE-PROD. We point the environment at CHA PRE-PROD which includes the completed bill run. That way the bill run will be correctly updated. We then use the updated information to create a fix migration.