Fix GitHub actions
Because we have had to cram in these large migrations and the sheer volume of them we've finally hit an issue in GitHub Actions.

```
ERROR: stdout maxBuffer length exceeded
Error: Process completed with exit code 1.
```

There is no options to increase the buffer size in Github actions or to tell db-migrate to shut up. So, we are left with trying to silence the output from `npm run migrate` by redirecting stdout to the black hole that is `/dev/null`.

We are only redirecting STDOUT so we should still see any errors that come out on STDERR.