Trying an alternate fix
That didn't work. But a quick Google of silent mode for db-migrate led us to this https://github.com/db-migrate/node-db-migrate/issues/453#issuecomment-278471581 which highlighted it's the boilerplate we use in the migrations causing the output.

So, we'll strip it in our new migration and if this works, will strip in from the others in a follow up PR.