Update /service-status version endpoints
https://eaflood.atlassian.net/browse/WATER-4096

Currently the `/service-status` endpoint in the UI is getting the versions for the various services from thier `/status` endpoints. As these endpoints are going to be updated to just return a static response of `{ "status": "alive" }`, this has been updated to `/health/info` which is an endpoint created for each repo a while ago to serve the repo's version.

This work is the first task in this issue: https://github.com/DEFRA/water-abstraction-team/issues/67