Returns version no. missing from `/service-status` (#2248)
https://eaflood.atlassian.net/browse/WATER-4096

After finishing work on the ticket to make the `/status` endpoints just return a static response. We have found that the `returns` version number is now missing from the UIs `/service-status` page.

In a previous PR https://github.com/DEFRA/water-abstraction-service/pull/2245 we had updated the endpoint that was being called to get the version number from `/status` to `/health/info` as we had updated the `/status` endpoint to just return a static response.

However, it would appear that for some reason the method used to get the version for the `returns` was different from all the others so was missed the first time round. This PR will fix that issue by ensuring that the `/health/info` endpoint is used when querying the `returns` service for its version.