Delete unused and unmaintained billing tests
https://github.com/DEFRA/water-abstraction-team/issues/70

Specifically, the folder `integration-tests/billing/tests`. We'd forgotten they existed and that we'd been advised to delete them by the previous team. Work on <https://github.com/DEFRA/water-abstraction-team/issues/25> required us to make changes to the files which brought them back to our attention again.

We now have up to date Cypress tests in [water-abstraction-acceptance-tests](https://github.com/DEFRA/water-abstraction-acceptance-tests) which do the same thing and the previous team admitted that they were not run or maintained. But they reticent to delete them as there may be info of value in the tests. However, we the new team will have little understanding of what's going on without trying to deep-dive through legacy code and we're all fed up of that.

So, let's reduce our codebase and avoid having to make changes to something that is dead to us in future!

Delete, delete, delete!