Improve Bill Run Queue Handling to Prevent Duplicate Jobs (#2293)
https://eaflood.atlassian.net/browse/WATER-4100

In our current codebase, we utilize BullMQ for queuing bill-run jobs, with each bill run assigned a unique ID that doubles as the job ID within the queue. BullMQ's behavior prevents the existence of two jobs with the same ID in the queue. Initially, we believed that double-clicking the "Send to Bill Run" button would result in the second request being ignored, as the subsequent job would share the same ID and BullMQ would naturally dismiss it.

However, upon closer examination, we realized that the job processes on the queue with remarkable speed, allowing it to be removed from the queue before the same job can re-enter when the user inadvertently double-clicks the button. This rapid processing has led to bill runs getting stuck, as the queue attempts to process the second request, which triggers a conflict request in the charging module (due to the bill having already been successfully billed). Consequently, the approveBatch function encounters an error, causing the status to revert from 'sent' to 'ready'.

This pull request addresses the issue by changing the configuration of the queue, specifically for the approveBatch queue.
This PR is changing the settings to maintain a record of completed bill-run jobs and remove any jobs that have failed. Keeping completed jobs in the queue allows BullMQ to identify duplicate bill runs and ignore them instead of processing them. Additionally, removing failed jobs enables the bill run to re-enter the queue if it fails without being flagged as a duplicate job.