Remove presroc supp. billing flag migration (#2300)
https://eaflood.atlassian.net/browse/WATER-4160

There has been a long-standing issue with the old scheme and the legacy code, which has resulted in large numbers of licences being flagged for presroc supplementary billing. There is no front-end way to remove billing flags, by design, to maintain an audit trail of licences that should be billed.

B&D have completed work to identify licences on a list provided by us where they have a billing flag but are billed to date. They've asked if we can remove the billing flag because there is no other way to do so.

Going forward we believe the charge version approval process is now the only thing that will put the presroc flag on. So, after this clean-up no licence _should_ get stuck with the flag on.

This change adds a migration to remove the flag from the licences B&D have specified.