Prevent error page showing for multiple return invitation requests (#2305)
https://eaflood.atlassian.net/browse/WATER-4104

This pull request aims to improve the handling of 'return invitation' notifications specifically that multiple requests are not being processed. Upon investigation, I can see that the legacy code already addresses multiple requests. The code examines the event's status, and if it doesn't match the 'processed' status, an error is raised.

This approach is beneficial as it guarantees that the service never sends duplicate notifications. However, it has a drawback: it displays an error page to the user, implying an issue with sending the batch, when, in fact, the batch still sends correctly. It's just that subsequent requests after the initial one aren't sent.

The changes introduced in this pull request involve an additional check. This new check examines if the event status is set to 'sending.' If this condition is met, instead of throwing an error, the system will return the data to the page and will not continue processing the request. This improvement ensures that users will see the correct page without encountering an error.

It's important to note that all other checks, including the status check, have been retained, as we want to account for any potential errors that might occur when the status is anything other than 'processing' or 'sending.'

There was no testing file for send-batch either so I have added this as well.