Improve Hand-Off Flow Notification Handling (#2295)
https://eaflood.atlassian.net/browse/WATER-4110

This pull request is working on preventing the system from sending multiple hand-off flow-level warning letters. Initially, it was assumed that each notification letter/email operated independently. Upon investigation, it was discovered that a single generic notification system was used, utilizing only one notification template with personalised data.

This constraint meant that a straightforward solution like restricting the BullMq queue to a single item wouldn't work. The data was also added to the database before job completion, initially marked as "draft," and subsequently updated to "sent" upon successfully sending the notification. This was done to maintain a record in case of any issues with the notify API, enabling the possibility of resending. In the table, there is a column called message_ref which is meant to indicate the type of notification being sent.

Due to the fact that all four notification types (1: Restriction, 2: Renewal letters, 3: Hands-off flow, 4: Resume) follow the same process, they are collectively labeled as 'notification' in the database message_ref column. To address this, a new column has been introduced in the scheduled_notifications table to record the notification type. Either 1,2,3 or 4.

To prevent the processing of multiple requests, we introduced a new "job_id" column within the "scheduled_notifications" table. Once a user inputs notification data and moves to the preview page, a unique ID is generated within the UI and gets passed to our service. When a notification request is received, this ID is utilized as the BullMQ job ID. If a duplicate request emerges, it will possess the same ID, prompting BullMQ to disregard it.

This solved the problem of multiple notifications being sent. However, we still encountered an issue where the record was added to the database for each request because this insertion occurred before the task was enqueued. To address this, we introduced a constraint on the "job_id" column, allowing only unique values to be inserted. As a result, any subsequent attempts to insert the same record are rejected by the database, ensuring uniqueness.