Fix email sending for non-notifications (#2352)
https://eaflood.atlassian.net/browse/WATER-4236

We recently made changes to try and stop multiple requests for the same task at the same time leading to duplicate jobs being added to the queue, for example, sending duplicate invites because someone managed to double-click the button in the UI.

To do that we came up with a solution whereby the UI generates a job ID for that user and session and then sends it in the payload. This in turn would be passed through to `enqueue()` in the second arg `options` as `jobId`.

Only in the land of notifications `options` isn't actually something to config things, but is the data itself handily named `n`!

So, we didn't see that in `src/modules/notify` and `src/modules/returns-notifications` (and possibly other places we don't know about) the second arg is used in an entirely different way.

> This is why consistency matters folks!

So, assuming this always gets sent through we switched to setting the `singletonKey:` to be it, which in turn helps form the unique job ID. Because it's not always sent through, this has led to jobs getting added with just the job name being set for the Job ID, for example `notify.send`.

This has led to erratic behaviour. For example, we have seen emails sending fine until we introduce an error, for example, requesting a password reset in a non-prod environment to an address that is not in Notify's whitelist. When this happens the job remains in the queue, being retried and eventually marked as errored. Meantime, subsequent requests to reset passwords or create new user accounts are blocked because BullMQ won't add a duplicate job to the queue.

We clearly don't have a good enough picture of all the different places emails are being sent and `enqueue()` and `scheduleSendEvent()` are being used. So, to avoid further issues until such time as we can rebuild this area of the service properly, this changes `scheduleSendEvent()` to use `rowJobId` if available, else revert to what it was doing before.