Fix - Delete licence with invalid leading space
https://eaflood.atlassian.net/browse/WATER-4357

It looks like the [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) has managed to import a duplicate NALD licence record (again!) this time because it differs from an existing one due to a leading space.

The Billing & Data team can clean up NALD by removing the problem record. But now it's been created in our DB the only way to remove it is with a migration. Until we remove it any searches where `NW/075/0012/011` appears in the results cause an error.