Fix - Delete licence with invalid leading space (#2430)
https://eaflood.atlassian.net/browse/WATER-4369

It looks like the water-abstraction-import has managed to import a duplicate NALD licence record (again!) this time because it differs from an existing one due to a leading space.

The Billing & Data team can clean up NALD by removing the problem record. But now it's been created in our DB the only way to remove it is with a migration. Until we remove it any searches where 7/34/06/*G/0309 appears in the results cause an error.