Stop marking bill runs as ready when they are not
https://eaflood.atlassian.net/browse/WATER-4345

We have been working replacing the legacy SROC annual billing engine with one in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). That engine is based on the one we built for supplementary billing which is also in **water-abstraction-system**.

We needed to deliver supplementary billing as soon as we can so only replaced part of the work the legacy service is doing. The new engines extract the transaction information and send it to the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). They even tell it generate the bill run. But they then send a request to this legacy service's `/water/1.0/billing/batches/{id}/refresh` endpoint. This does the job of querying the CHA for the updated invoice and transaction information after the bill run has been generated. It then updates the WRLS records based on this.

At the time it was deemed out of scope to allow us to focus on getting the bill run generated. But what we've learned over time is that this project is immediately marking the bill run as `READY` once it sees the CHA has finished generating before it has finished updating itself.

Those of us who have access to the logs can see the service is still busy working. Users of the service will experience this is they compare a download of the bill run generated straight away to one generated later.

We believe this confusion is contributing to the problems we experience during annual billing. For example, users may believe it is fine to start the next bill run unaware this project is busy hammering the CHA for billing information.

So, this change stops the legacy service marking the bill run as ready _until_ the 'update-invoices' process is complete. It will make the billing process appear to take longer but it at least will be honest and hopefully reduce the issues users experience.