Replace child-process based code with standard
The function we are replacing was based on `src/modules/billing/jobs/update-invoices.js` spinning up a child process for this module.

Then it would start the process by calling `child.send()`. Communication between the main process and the child process happens using `message()` events.

It's all very complex and no doubt was intended to improve performance. But this job already wasn't blocking anything. It was purely the fact it was called from a job which by default expects to be completed in no more than 30 secs where the wheels must have started falling off.

The moment it was spawned into a child process keeping track of what was going on, specifically when they went wrong stopped working. Hence bill runs never getting updated with no logs or errors given.

We'll no longer be forking a child process to run this code so we can strip out that element. That also means we can do things like pass the logger so we can start seeing log output again.

The other thing the old code was doing was basically 'attacking' the CHA. If running this process for an Anglian bill run of 2500 invoices this code was firing 2500 requests at the charge module as fast as it could. What this results in is slowing down how long each request takes to process whilst increasing the chance of an error occurring.

So, we use a pattern we've implemented in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) for batching processes. This way the CHA is only being asked to process 10 at a time giving it more resource to manage each request and reducing the chance for request to get dropped.

Other than that, we have left the process as it was. Mad as some of the code is if we were going to re-write, we may as well do it properly in our own code base.