Update the update-invoices worker
The key change is dropping the need to use Node's child-process to generate a fork to run the update invoices process in.

The job now follows a more standard pattern hence we bring in `onComplete` and `onFailed`.

This also means we are no longer relying on some of the generic job modules the previous team created so we can remove those dependencies.

Finally, the job was setup to retry up to 10 times with an exponential back off in the event of failure. It sounds like a great idea but in reality it would lead to bill runs being 'processed' for exceedingly long times. And as we've learnt, beyond the odd request dropping now and then, if the job failed it was something that retrying wouldn't solve so was pointless.

So, we remove all job retry config as well.