Ensure update-invoices sets bill run to ready (#2445)
https://eaflood.atlassian.net/browse/WATER-4345

We have been working on replacing the legacy SROC annual billing engine with one in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). That engine is based on the one we built for supplementary billing, also in **water-abstraction-system**.

In the previous change, we updated this project to [Stop marking bill runs as ready when they are not](https://github.com/DEFRA/water-abstraction-service/pull/2442).

The refresh job was marking the bill run as ready and _then_ kicking off the process to refresh WRLS with data from the Charging Module API.

As part of that, we also tried to make the process more robust and stop depending on creating a child process.

All seemed well until we tested with our largest region. The bill run was processed and all the invoices were updated, but the `onComplete` never fired, meaning the bill run never got out of 'building'.

So, to avoid that happening again we have

- moved the status setting logic into the man handler rather than relying on the `onComplete` and `onFailed` event handlers (we don't trust them!)
- updated the worker options to match other jobs that need a very long time to complete

We suspect that the previous config meant the job timed out so it was pulled and the `onComplete` never fired. But this wouldn't stop the process from continuing.