Implement the hack
We copy the existing `update-invoices.js` and then simply change the job name and the status it sets on the bill run when complete.

We tweak both the `refresh-totals.job` and its helper service `cm-fresh-service.js` to allow us to pick either `update-invoices.js` or our new `update-invoice-references.js` as the next job depending on whether we are generating the bill run or sending it.

After that we let the insanity commence as it was doing before!