Fix refresh totals job
https://eaflood.atlassian.net/browse/WATER-4345

In [Fix broken send bill run functionality](https://github.com/DEFRA/water-abstraction-service/pull/2446) we removed some config from the `src/modules/billing/jobs/refresh-totals.js` job that allowed it to be retried within BullMQ.

We figured this exponential retry is why the job gets stuck in the background. But actually we were just displaying our ignorance. Because it can take a moment for the CHA to complete the send process it is likely that it won't have a status of `billed` if immediately queried after sending it.

It turns out the previous team used the BullMQ mechanism for retrying to get the CHA bill run summary rather than their own for this job. So, removing the config broke the process. We didn't spot this until put in one of our AWS environments where the CHA takes a little longer than it does locally.

This change puts the config back.