Fix remove bill not flagging licences for supp.
https://eaflood.atlassian.net/browse/WATER-4416

We spotted that our new billing views had dropped the button that allows a user to remove a licence from a bill run (only applies to `READY` bill runs and bills with more than one licence).

We put the button back in and even gave the confirmation page a refresh. But we're not in a position to migrate the actual work yet. So, our page like the old one pings the same legacy endpoint as [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui).

Our QA team being as conscientious as always double-checked all aspects, following up on the message that both the old and new pages display stating the licence will be flagged for the next supplementary bill run. They found it always flags them for PRESROC, when it should have been flagging some for SROC supplementary billing.

Which is where this fix comes in. We display the same message on confirm remove bill page. Only this time testing found the legacy code doesn't flag anything at all! 😩🤦

This change fixes the endpoint so now when its pinged it will flag the licences correctly. If the bill run the bill is being removed from is SROC then all licences in the bill need to be flagged for the next SROC supplementary bill run. If the bill run is PRESROC, then they need to be flagged for the next PRESROC supplementary bill run.