Enhance FixtureLoader
To support our billing tests we want to be able to create a 'sent' annual bill run in either the current year or the previous year. We don't want to hard code the start and end year for the bill run record as it'll just break once we move into 2025.

So, it needs to be dynamic. By looking at some of the existing fixtures we've figured out the mechanism for getting the loader to generate and insert dynamic values into the yaml files.

First step is to enhance the loader, specifically `calcValue()` to provide current and previous financial year values.