Remove licence-not-in-charge-version-workflow job
**Context**

This service has the scheduled job [Licence not in charge version workflow](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#licence-not-in-charge-version-workflow). It's intent is to add a `water.charge_version_workflow` record any time it sees a new or updated licence.

Our understanding is this is intended to keep the licence from being included in a bill run until Billing & Data has had a chance to review its charge versions.

We have had a call logged with us because this job doesn't appear to be running any more. Billing & Data know they have made changes in the upstream service NALD which should have resulted in licences being added to workflow but hasn't.

Our digging into the logs has found this job runs sporadically at best! There have been periods spanning weeks where the job has not run, even though it is scheduled to run 4 times a day! We believe the fact it is annual billing time has led to a renewed focus and so the team have only just spotted something is not right.

Why the job is being sporadic we cannot answer. The job is running using [BullMQ](https://docs.bullmq.io/) message queue architecture. In this case, the schedule is set using config passed to BullMQ when creating the [repeatable job](https://docs.bullmq.io/guide/jobs/repeatable). Everything appears to be correct. But we've seen this behaviour before and we know we are now 4 versions behind the current version of BullMQ.

Not wishing to plough more time and resources into confirming exactly _why_ BullMQ is being flaky we are instead using this as an opportunity to migrate the job away from the legacy code.

This means the job is being replaced by [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/903).

So, this change strips the existing job from the project to avoid any issues and remove the redundant code.